requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
aiosqlite==0.20.0
httpx==0.27.2
binance-connector==3.8.0
python-telegram-bot==21.6
Jinja2==3.1.4
